src-tauri/python/requirements.txt
# Python ML dependencies for pose estimation and audio processing
# Install with: pip install -r requirements.txt

# MediaPipe for pose, face, and hand tracking
mediapipe>=0.10.0

# OpenAI Whisper for speech transcription
openai-whisper>=20230314

# pyannote.audio for speaker diarization
pyannote.audio>=3.1.0

# SpeechBrain for emotion recognition
speechbrain>=0.5.16

# Audio processing
librosa>=0.10.0
soundfile>=0.12.0
torchaudio>=2.0.0

# Core ML dependencies
torch>=2.0.0
numpy>=1.24.0
opencv-python>=4.8.0

# Optional: GPU acceleration (uncomment if you have CUDA)
# torch>=2.0.0+cu118 --extra-index-url https://download.pytorch.org/whl/cu118
